+++ src/protocol/dap.cpp
@@ -947,7 +947,7 @@ static std::string ReadData(std::istream &cin)
 
             char *p = nullptr; // NOLINT(misc-const-correctness)
             errno = 0;
-            content_len = strtoul(&line[CONTENT_LENGTH.size()], &p, 10);
+            content_len = static_cast<long>(strtoul(&line[CONTENT_LENGTH.size()], &p, 10));
             if (errno == ERANGE || (*p != 0 && (isspace(*p) == 0)))
             {
                 LOGE("protocol violation: '%s'", line.c_str());
